move go build error parsing
